fix) fixed docker-compose file for a better container handling
1. services.consulserver.environment
These variables are used to tell consul to bind itself to the host local network

2. services.consulserver.volumes
This setting will allow consul to store consul related data on host's disk. This will save the current status and will secure the current setting even after a system reboot

3. services.consulserver.command
This command runs consul to bind itself to the eth0 interface, save the current settings at /tmp/consul, enables ui and so on

4. services.configserver.environment
Loads sensitive information from .env file. This file is in .gitignore and must be served manually and should be hidden from the public

5. services.configserver.links
It enables this container to connect to consul discovery using http://consulserver:8500 rather than the docker container ip address