fix) Changed dockerfile
Base image: openjdk:8-jdk-alpine

What does it do?
1. Downloads Oracle jce_policy for a stronger encryption method.
2. Unpack the fatjar as described here:
https://spring.io/guides/gs/spring-boot-docker/

"...This Dockerfile has a DEPENDENCY parameter pointing to a directory where we have unpacked the fat jar. If we get that right, it already contains a BOOT-INF/lib directory with the dependency jars in it, and a BOOT-INF/classes directory with the application classes in it. Notice that we are using the application’s own main class hello.Application (this is faster than using the indirection provided by the fat jar launcher).
..."